updated BaseUrlTest to use HttpKernelBrowser when possible, handling all possible deprecations for TestClient
