+++ tests/helpers/AbstractClientForClient.php
@@ -1,0 +1,6 @@
+<?php
+
+
+namespace Behat\Mink\Tests\Driver;
+
+abstract class AbstractClient extends \Symfony\Component\BrowserKit\Client {}
