+++ tests/helpers/DoRequest5.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Behat\Mink\Tests\Driver;
+
+require_once __DIR__ . '/AbstractTestClient.php';
+
+abstract class DoRequestReturnObject extends AbstractTestClient
+{
+    protected function doRequest($request)
+    {
+        return $this->_doRequest($request);
+    }
+}
