+++ tests/helpers/DoRequest72.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Behat\Mink\Tests\Driver;
+
+require_once __DIR__ . '/AbstractTestClient.php';
+
+abstract class DoRequestReturnObject extends AbstractTestClient
+{
+    protected function doRequest($request): object
+    {
+        return $this->_doRequest($request);
+    }
+}
